pyproject.toml
[project]
name = "saya-ai-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "crewai>=0.186.1",
    "fastapi>=0.116.1",
    "langchain-openai>=0.3.33",
    "openai>=1.107.1",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
    "sqlalchemy>=2.0.0",
    "pymysql>=1.1.0",
    "cryptography>=41.0.0",
    "cohere>=5.0.0",
    "pypdf2>=3.0.0",
    "langchain>=0.1.0",
    "langchain-text-splitters>=0.0.1",
    "httpx>=0.25.0",
    "tidb-vector>=0.0.9",
]
